requirements.txt
aiohttp
websockets
pandas
requests
python-dotenv
psycopg2-binary